pyproject.toml
[tool.poetry]
name = "projectwithpoetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
python-decouple = "^3.8"
pyjwt = "^2.8.0"
python-jose = "^3.3.0"
boto3 = "^1.34.95"
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.7.1"}
requests = "^2.31.0"
uuid = "^1.30"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
